refactor: move consentStatus + use expired status in request header
