+++ src/features/amUI/consent/ConsentRequestPage/ConsentRequestPage.module.css
@@ -11,6 +11,10 @@
   margin-bottom: var(--ds-size-12);
 }
 .headerBlock {
+  display: inline-flex;
+  align-items: center;
+  flex-wrap: wrap;
+  gap: var(--ds-size-6);
   padding-top: var(--ds-size-14);
   margin-bottom: 3px;
 }
